requirements.txt
numpy==1.23.5
Pillow==9.3.0
PyQt6==6.4.0
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
Unidecode~=1.3.6